requirements.txt
# API & Data Fetching
requests==2.31.0
alpha-vantage==2.3.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Processing & Analysis
pandas==2.1.3
numpy==1.26.2

# Airflow
apache-airflow==2.7.1

# Dashboard
streamlit==1.28.2
plotly==5.18.0

# Utilities
python-dotenv==1.0.0